+++ packages/fighting-design/header/src/header.vue
@@ -1,7 +1,7 @@
 <script lang="ts" setup>
   import { Props } from './header'
   import { computed } from 'vue'
-  import { isString } from '../../_utils'
+  import { sizeChange } from '../../_utils'
   import type { CSSProperties, ComputedRef } from 'vue'
 
   const prop = defineProps(Props)
@@ -10,8 +10,8 @@
     const { height, padding } = prop
 
     return {
-      height: isString(height) ? height : height + 'px',
-      padding: isString(padding) ? padding : padding + 'px'
+      height: sizeChange(height),
+      padding: sizeChange(padding)
     } as const
   })
 </script>