+++ packages/fighting-design/toolbar/src/toolbar.vue
@@ -37,7 +37,6 @@
     return style
   })
 
-  // const handleClick: b = (evt: MouseEvent & { path: HTMLElement[] }): void => {
   const handleClick: b = (evt: MouseEvent): void => {
     // 针对 Safari 点击事件 event.path 为 undefined 不存在做兼容
     // https://blog.csdn.net/weixin_43388844/article/details/120002440